completed websocket implementation... testing to be done and media handlers are still not creating. expecting frontend start after testing

+++ dbhandler/solar_msg_control.py
@@ -2,7 +2,9 @@ import uuid
 import objects
 from error_handler import ErrorCodes
 from objects import SolarMessage
+
 from dbhandler import sql_handler
+from dbhandler.solar_control import SolarControl
 
 error_codes = ErrorCodes()
 
@@ -10,15 +12,20 @@ class SolarMsgControl:
 
     def add_message(sl_id, user_id, message, attributes):
         msg_id = str(uuid.uuid4())
-        query = "insert into SolarMessages (msg_id, sl_id, id, data, attributes) values (%s, %s, %s, %s, %s)"
-        params = (msg_id, sl_id, user_id, message, attributes)
-        result = sql_handler.put_query(query, params)
-        if result is None:
-            return error_codes.DB_ERROR
+        if SolarControl.is_solar_member(sl_id, user_id) == 1:
+            query = "insert into SolarMessages (msg_id, sl_id, id, data, attributes) values (%s, %s, %s, %s, %s)"
+            params = (msg_id, sl_id, user_id, message, attributes)
+            result = sql_handler.put_query(query, params)
+            if result is None:
+                return error_codes.DB_ERROR
+            else:
+                return 0
         else:
-            return 0
+            return error_codes.USER_NOT_SOLAR_MEMBER
 
-    def update_message(msg_id, message):
+    def update_message(id, msg_id, message):
+        if id != (sql_handler.put_query("select id from solarmessages where msg_id = %s", (msg_id))[0][0]):
+            return error_codes.UNAUTHORISED_REQUEST
         query = "update SolarMessages set data = %s, edited = edited + 1 where msg_id = %s"
         params = (message, msg_id)
         result = sql_handler.put_query(query, params)
@@ -27,19 +34,27 @@ class SolarMsgControl:
         else:
             return 0
 
-    def get_messages(sl_id, offset=0, limit=100):
-        query = "select * from SolarMessages where sl_id = %s order by at desc limit %s offset %s"
-        params = (sl_id, limit, offset)
-        result = sql_handler.put_query(query, params)
-        if result is None:
-            return error_codes.DB_ERROR
+    def get_messages(sl_id, id, offset=0, limit=100):
+        if SolarControl.is_solar_member(sl_id, id) == 1:
+            query = "select * from SolarMessages where sl_id = %s order by at desc limit %s offset %s"
+            params = (sl_id, limit, offset)
+            result = sql_handler.put_query(query, params)
+            if result is None:
+                return error_codes.DB_ERROR
+            else:
+                list_messages = []
+                for message_data in result:
+                    list_messages.append(SolarMessage(*message_data))
+                return list_messages
         else:
-            list_messages = []
-            for message_data in result:
-                list_messages.append(SolarMessage(*message_data))
-            return list_messages
+            return error_codes.UNAUTHORISED_REQUEST
 
-    def delete_message(msg_id):
+    def delete_message(msg_id, id):
+        req = sql_handler.put_query("select id from solarmessages where msg_id = %s", (msg_id))
+        if req == []:
+            return error_codes.MESSAGE_DOES_NOT_EXIST
+        if id != req[0][0]:
+            return error_codes.UNAUTHORISED_REQUEST
         query = "delete from SolarMessages where msg_id = %s"
         params = (msg_id,)
         result = sql_handler.put_query(query, params)
